Move svgr and dynamic-imports to cra setup
